+++ routes/authRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const authController = require('../controllers/authController.js');
+
+const router = express.Router();
+
+router.post('/login', authController.loginUser);
+
+module.exports = router;
