+++ routes/imageRoutes.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const imageController = require('../controllers/imageController');
+const { verifyToken } = require('../middlewares/auth');
+
+const router = express.Router();
+
+router.post('/create', verifyToken, imageController.createImage);
+router.get('/list', verifyToken, imageController.listImages);
+
+module.exports = router;
